Added a sample CombatLog
Warning! This CombatLog is huge: it contains three matches and the time
between the matches, though not much was done during those times. It is
one of my own CombatLogs. It is only a sample, testing CombatLogs must
be shorter to keep things understandable for humans. Picking a few lines
from this CombatLog is fine.