Merge ui_testing into development to add UI unittests for #46 (#64)
* Added basic UI testing for #46

* Update OS to have logged in user and UI

* Removed os: unstable

It did not make any difference to the resulting output of AppVeyor

* Added cp assets for build artifact

Otherwise the assets are not included in the build artifacts

* Added three debug messages for AppVeyor build

The AppVeyor build hangs on the command "python testing.py" currently
with no more output than just four dots "....". I would like to know
where exactly it hangs on.

* Add RDP for stuck session investigation

* Added makedirs for AppVeyor

And remove the debug messages that were not of any help. Using the RDP
session, I was able to determine that "tkMessageBox.showerror" was
called because of the non-existent CombatLogs folder. However, using
`mock.patch("tkMessageBox.showerror", return_value=0)` did not have any
effect.

* Updated with messagebox replacement

* Added file and test for file

* Added the recursive parameter to the cp command