Merge codecov into development to add codecov functionality (#65)
* Added codecov to Travis and AppVeyor

* Updated with coverage

* Added codecov badge

* Updated coverage command

* Updated codecov command

Apparently the folder %PYTHON%\\Scripts is added to the PATH environment
variable on AppVeyor Python 2.7 x86 Virtual Machines, but this is not
the case on the x64 variant. Therefore, the path supplied needs to
absolute and not relative.

* Updated path for codecov with Scripts

* Updated with time.sleep

This is to prevent "ttk::ThemeChanged" errors, which sometimes occur in
the TestUI class during testing on AppVeyor. I think this is caused by
the first window not closing quickly enough.