Improved UI and added basic sorting for #37
While the sorting does not yet work correctly for damage taken
(magnitude) or the enemies' ID (alphabetical), it does work for the
damage dealt part. Also the headings are now shown and the obsolete
alternate colouring is now gone. Selecting an item in a Treeview with
the plastik theme looks quite good.