Updated sceen.py
Data format for the storage of the screen parsing data is provided in a
docstring. cPickle is a fast method to store all kinds of Python
objects, including dictionaries. This might turn to be a very large file
if the function is used a lot. An SQLite database would be overkill in
my opinion.