Update method of determining DPI scaling
There is a bug pywin32, specifically one described here: https://sourceforge.net/p/pywin32/bugs/609/ . It causes the program not to exit correctly if win32ui is imported. As a work-around, Overlay now first attempts to get its DPI scaling value from a master widget. If this is not possible, then it will use win32ui, accepting that it will not be able to exit correctly.