Add pseudo-code for new parsing engine
A whole new parsing engine would be a good idea, I think, because the new one is quite old and adapting it for #147 for example would be a nightmare. Having a single parsing engine for both realtime and file parsing would make things easier to maintain. Code is going to be at least partially based on realtime.py.