Add snaptoplevel.py for SnapToplevel
SnapToplevel is currently not yet available from the ttkwidgets packages, as it is not yet suitable to use on all operating systems. Therefore, as the author, I have copied it over to the /widgets folder of the repository to allow importing it without having to manually merge snaptoplevel into master on the ttkwidgets repository and installing that code.