Re-add Travis-CI badge to README
As the project grows larger and larger, automated tests are definitely required to make sure that nothing big is broken in the code-base when making small changes. With the newly added native Linux support, the parsing functions should be perfectly testable on Linux as well. Testing the UI and user interaction is difficult, as it's not what Tkinter was made for and the program depends highly on the presence of SWTOR.